move FinanceService to new module
